Move `SessionAverageHitErrorTracker` to more general namespace
